pyproject.toml
[project]
name = "datacrawler"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler==3.10.4",
    "beautifulsoup4==4.12.3",
    "emoji==2.14.0",
    "httpx==0.27.2",
    "lxml==5.3.0",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "pydantic==2.9.2",
    "python-dotenv==1.0.1",
    "pyyaml==6.0.2",
    "readability-lxml==0.8.1",
    "regex==2024.11.6",
    "selenium>=4.39.0",
    "sqlalchemy>=2.0.45",
    "tldextract==5.1.2",
    "webdriver-manager>=4.0.2",
]
